+++ lib/virtualbadge/view/draw_badge.dart
